[Refactor][OPS] split tileops/ops/elementwise.py into per-cluster package (#1388)
Closes #1378

## Summary

- Replaced 3415-line monolith `tileops/ops/elementwise.py` with a
`tileops/ops/elementwise/` package, one module per cluster (16 cluster
modules + `_base.py`).
- Umbrella classes `UnaryOp`, `BinaryOp`, `FusedGatedOp` moved to
`_base.py`; `__init__.py` re-exports every previously-public symbol via
explicit `__all__`.
- Pure file-layout move: no class signature, docstring, or
kernel-binding changes; all 295 elementwise tests pass; no file-level
lint suppressions introduced.
- Pyright shows pre-existing typing issues in the relocated code that
were carried over verbatim from the original file — out of scope to fix
in this refactor.

## Test plan

- [x] AC-1: Modified files pass unit tests (pytest tests/ops/).
- [x] AC-2: tileops/ops/elementwise.py no longer exists;
tileops/ops/elementwise/ is a package with __init__.py.
- [x] AC-3: python -c "from tileops.ops.elementwise import UnaryOp,
BinaryOp, FusedGatedOp" succeeds.
- [x] AC-4: python -c "from tileops.ops.elementwise import
SiluAndMulFwdOp, ReluFwdOp, AddFwdOp, ExpFwdOp, DivFwdOp" succeeds.
- [x] AC-5: tests/ops/test_elementwise_fp8.py passes without
modification to its imports.
- [x] AC-6: grep -nE '^# ruff:|^# flake8:' tileops/ops/elementwise/
returns nothing.
- [x] AC-7: External tracker flip — out of scope.

---------
